// src/types/theme.ts
export type Theme = {
  mode: 'light' | 'dark' | 'system';
  colors: {
    theme: string;
    appBg: string;
    appBgLight: string;
    background: string;
    text: string;
    primary: string;
    secondary: string;
    border: string;
    placeholder: string;
    card: string;
    discount: string;
    accent: string;
    veg: string;
    error: string;
    textSecondary: string;
    divider: string;
    success: string;
  };
};